feat(spider-label): 自定义注册spide-label
